[EST-98] Refactor : common 내 constant 패키지로 이동
