--- src/main/java/com/blackshoe/esthete/common/Gender.java
@@ -1,5 +1,0 @@
-package com.blackshoe.esthete.common;
-
-public enum Gender {
-    MALE, FEMALE
-}
