+++ src/main/java/com/blackshoe/esthete/common/constant/Gender.java
@@ -1,0 +1,5 @@
+package com.blackshoe.esthete.common.constant;
+
+public enum Gender {
+    MALE, FEMALE
+}
